% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_turo_fonts.R, R/install_fonts.R
\name{import_turo_fonts}
\alias{import_turo_fonts}
\title{Import Turo fonts}
\usage{
import_turo_fonts()

import_turo_fonts()
}
\description{
Download the Turo fonts from github et import them in R envr

Install the turo fonts
}
\details{
Import Turo fonts
}
\examples{
#import_turo_fonts()
import_turo_fonts()
}
\author{
Toma Gulea
}

