% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_tech.R
\name{scale_color_tech}
\alias{scale_color_tech}
\title{Tech scale color for ggplot2}
\usage{
scale_color_tech(theme = "airbnb", tech_key = list(airbnb = c("#FF5A5F",
  "#FFB400", "#007A87", "#FFAA91", "#7B0051"), facebook = c("#3b5998",
  "#6d84b4", "#afbdd4", "#d8dfea"), google = c("#5380E4", "#E12A3C", "#FFBF03",
  "#00B723"), etsy = c("#F14000", "#67B6C3", "#F0DA47", "#EBEBE6", "#D0D0CB"),
  twitter = c("#55ACEE", "#292f33", "#8899a6", "#e1e8ed"), X23andme =
  c("#3595D6", "#92C746", "#F2C100", "#FF6D19", "#6F3598")))
}
\description{
tech scale color
}
\examples{
scale_color_tech(theme="etsy")
}
\author{
Ricardo Bion
}
