% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale_fill_tech.R
\name{scale_fill_tech}
\alias{scale_fill_tech}
\title{Tech scale fill for ggplot2}
\usage{
scale_fill_tech(theme = "airbnb")
}
\description{
tech scale fill
}
\examples{
scale_fill_tech(theme="etsy")
}
\author{
Ricardo Bion
}

