% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_airbnb_fancy.R
\name{theme_airbnb_fancy}
\alias{theme_airbnb_fancy}
\title{Airbnb theme for ggplo2}
\usage{
theme_airbnb_fancy(base_size = 2.5)
}
\value{
A theme element
}
\description{
Import Airbnb ggplot2 theme. It builds on theme_classic. You can change font family or size.
}
\examples{
qplot(1) + theme_airbnb_fancy()
}
