% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_turo_fancy.R
\name{theme_turo_fancy}
\alias{theme_turo_fancy}
\title{turo theme for ggplo2}
\usage{
theme_turo_fancy(base_size = 2.5)
}
\value{
A theme element
}
\description{
Import turo ggplot2 theme. It builds on theme_classic. You can change font family or size.
}
\examples{
qplot(1) + theme_turo_fancy()
}

