% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tech.R
\name{techGrob}
\alias{techGrob}
\title{Tech geom}
\usage{
techGrob(x, y, size, theme, geom_key = list(airbnb =
  "http://logok.org/wp-content/uploads/2014/07/airbnb-logo-belo-880x628.png",
  etsy =
  "https://d3rnbxvnd0hlox.cloudfront.net/images/channels/64/icons/large.png",
  facebook =
  "http://blog.icracked.com/wp-content/uploads/2015/03/facebook-icon.png",
  twitter =
  "https://lh3.ggpht.com/lSLM0xhCA1RZOwaQcjhlwmsvaIQYaP3c5qbDKCgLALhydrgExnaSKZdGa8S3YtRuVA=w300",
  google =
  "http://logok.org/wp-content/uploads/2015/09/Google-logo-2015-G-icon.png"))
}
\description{
Tech geom
}
\author{
Ricardo Bion
}

